Move events code to openai.py
