+++ logfire/_internal/integrations/llm_providers/openai.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import contextlib
 from typing import TYPE_CHECKING, Any, cast
 
 import openai
@@ -9,10 +10,13 @@ from openai.types.chat.chat_completion_chunk import ChatCompletionChunk
 from openai.types.completion import Completion
 from openai.types.create_embedding_response import CreateEmbeddingResponse
 from openai.types.images_response import ImagesResponse
+from openai.types.responses import Response
 from opentelemetry.sdk.trace import ReadableSpan
 from opentelemetry.trace import get_current_span
 
-from ...utils import handle_internal_errors
+from logfire import LogfireSpan
+
+from ...utils import handle_internal_errors, log_internal_error
 from .types import EndpointConfig, StreamState
 
 if TYPE_CHECKING:
@@ -182,3 +186,86 @@ def is_async_client(client: type[openai.OpenAI] | type[openai.AsyncOpenAI]):
         return False
     assert issubclass(client, openai.AsyncOpenAI), f'Expected OpenAI or AsyncOpenAI type, got: {client}'
     return True
+
+
+@handle_internal_errors
+def get_responses_api_message_events(response: Response | None, inputs: str | list[dict[str, Any]] | None = None):
+    events: list[dict[str, Any]] = []
+    tool_call_id_to_name: dict[str, str] = {}
+    if response and (instructions := getattr(response, 'instructions', None)):
+        events += [
+            {
+                'event.name': 'gen_ai.system.message',
+                'content': instructions,
+                'role': 'system',
+            }
+        ]
+    if inputs:
+        if isinstance(inputs, str):  # pragma: no cover
+            inputs = [{'role': 'user', 'content': inputs}]
+        for inp in inputs:
+            events += input_to_events(inp, tool_call_id_to_name)
+    if response and response.output:
+        for out in response.output:
+            for message in input_to_events(out.model_dump(), tool_call_id_to_name):
+                events.append({**message, 'role': 'assistant'})
+    return events
+
+
+def input_to_events(inp: dict[str, Any], tool_call_id_to_name: dict[str, str]):
+    try:
+        events: list[dict[str, Any]] = []
+        role: str | None = inp.get('role')
+        typ = inp.get('type')
+        content = inp.get('content')
+        if role and typ in (None, 'message') and content:
+            event_name = f'gen_ai.{role}.message'
+            if isinstance(content, str):
+                events.append({'event.name': event_name, 'content': content, 'role': role})
+            else:
+                for content_item in content:
+                    with contextlib.suppress(KeyError):
+                        if content_item['type'] == 'output_text':  # pragma: no branch
+                            events.append({'event.name': event_name, 'content': content_item['text'], 'role': role})
+                            continue
+                    events.append(unknown_event(content_item))  # pragma: no cover
+        elif typ == 'function_call':
+            tool_call_id_to_name[inp['call_id']] = inp['name']
+            events.append(
+                {
+                    'event.name': 'gen_ai.assistant.message',
+                    'role': 'assistant',
+                    'tool_calls': [
+                        {
+                            'id': inp['call_id'],
+                            'type': 'function',
+                            'function': {'name': inp['name'], 'arguments': inp['arguments']},
+                        },
+                    ],
+                }
+            )
+        elif typ == 'function_call_output':
+            events.append(
+                {
+                    'event.name': 'gen_ai.tool.message',
+                    'role': 'tool',
+                    'id': inp['call_id'],
+                    'content': inp['output'],
+                    'name': tool_call_id_to_name.get(inp['call_id'], inp.get('name', 'unknown')),
+                }
+            )
+        else:
+            events.append(unknown_event(inp))
+        return events
+    except Exception:  # pragma: no cover
+        log_internal_error()
+        return [unknown_event(inp)]
+
+
+def unknown_event(inp: dict[str, Any]):
+    return {
+        'event.name': 'gen_ai.unknown',
+        'role': inp.get('role') or 'unknown',
+        'content': f'{inp.get("type")}\n\nSee JSON for details',
+        'data': inp,
+    }
