+++ logfire/_internal/integrations/openai_agents.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import contextlib
 import contextvars
 import inspect
 import sys
@@ -36,6 +35,7 @@ from opentelemetry.trace import NonRecordingSpan, use_span
 from typing_extensions import Self
 
 from logfire._internal.formatter import logfire_format
+from logfire._internal.integrations.llm_providers.openai import get_responses_api_message_events
 from logfire._internal.scrubbing import NOOP_SCRUBBER
 from logfire._internal.utils import handle_internal_errors, log_internal_error, truncate_string
 
@@ -403,87 +403,7 @@ def get_magic_response_attributes() -> dict[str, Any]:
         return {}
 
 
-@handle_internal_errors
 def get_response_span_events(span: ResponseSpanData):
-    events: list[dict[str, Any]] = []
     response = span.response
-    inputs = span.input
-    tool_call_id_to_name: dict[str, str] = {}
-    if response and (instructions := getattr(response, 'instructions', None)):
-        events += [
-            {
-                'event.name': 'gen_ai.system.message',
-                'content': instructions,
-                'role': 'system',
-            }
-        ]
-    if inputs:
-        if isinstance(inputs, str):  # pragma: no cover
-            inputs = [{'role': 'user', 'content': inputs}]
-        for inp in inputs:  # type: ignore
-            inp: dict[str, Any]
-            events += input_to_events(inp, tool_call_id_to_name)
-    if response and response.output:
-        for out in response.output:
-            for message in input_to_events(out.model_dump(), tool_call_id_to_name):
-                events.append({**message, 'role': 'assistant'})
-    return events
-
-
-def input_to_events(inp: dict[str, Any], tool_call_id_to_name: dict[str, str]):
-    try:
-        events: list[dict[str, Any]] = []
-        role: str | None = inp.get('role')
-        typ = inp.get('type')
-        content = inp.get('content')
-        if role and typ in (None, 'message') and content:
-            event_name = f'gen_ai.{role}.message'
-            if isinstance(content, str):
-                events.append({'event.name': event_name, 'content': content, 'role': role})
-            else:
-                for content_item in content:
-                    with contextlib.suppress(KeyError):
-                        if content_item['type'] == 'output_text':  # pragma: no branch
-                            events.append({'event.name': event_name, 'content': content_item['text'], 'role': role})
-                            continue
-                    events.append(unknown_event(content_item))  # pragma: no cover
-        elif typ == 'function_call':
-            tool_call_id_to_name[inp['call_id']] = inp['name']
-            events.append(
-                {
-                    'event.name': 'gen_ai.assistant.message',
-                    'role': 'assistant',
-                    'tool_calls': [
-                        {
-                            'id': inp['call_id'],
-                            'type': 'function',
-                            'function': {'name': inp['name'], 'arguments': inp['arguments']},
-                        },
-                    ],
-                }
-            )
-        elif typ == 'function_call_output':
-            events.append(
-                {
-                    'event.name': 'gen_ai.tool.message',
-                    'role': 'tool',
-                    'id': inp['call_id'],
-                    'content': inp['output'],
-                    'name': tool_call_id_to_name.get(inp['call_id'], inp.get('name', 'unknown')),
-                }
-            )
-        else:
-            events.append(unknown_event(inp))
-        return events
-    except Exception:  # pragma: no cover
-        log_internal_error()
-        return [unknown_event(inp)]
-
-
-def unknown_event(inp: dict[str, Any]):
-    return {
-        'event.name': 'gen_ai.unknown',
-        'role': inp.get('role') or 'unknown',
-        'content': f'{inp.get("type")}\n\nSee JSON for details',
-        'data': inp,
-    }
+    inputs: str | list[dict[str, Any]] | None = span.input  # type: ignore
+    return get_responses_api_message_events(response, inputs)
